+++ src/Stack/Types/Config.hs
@@ -193,11 +193,12 @@ data ExecOptsExtra
         }
 -- | Parsed global command-line options.
 data GlobalOpts = GlobalOpts
-    { globalLogLevel     :: LogLevel -- ^ Log level
-    , globalConfigMonoid :: ConfigMonoid -- ^ Config monoid, for passing into 'loadConfig'
-    , globalResolver     :: Maybe AbstractResolver -- ^ Resolver override
-    , globalTerminal     :: Bool -- ^ We're in a terminal?
-    , globalStackYaml    :: Maybe FilePath -- ^ Override project stack.yaml
+    { globalReExec       :: !Bool
+    , globalLogLevel     :: !LogLevel -- ^ Log level
+    , globalConfigMonoid :: !ConfigMonoid -- ^ Config monoid, for passing into 'loadConfig'
+    , globalResolver     :: !(Maybe AbstractResolver) -- ^ Resolver override
+    , globalTerminal     :: !Bool -- ^ We're in a terminal?
+    , globalStackYaml    :: !(Maybe FilePath) -- ^ Override project stack.yaml
     } deriving (Show)
 
 -- | Either an actual resolver value, or an abstract description of one (e.g.,
@@ -262,16 +263,11 @@ instance HasEnvConfig EnvConfig where
 data LoadConfig m = LoadConfig
     { lcConfig          :: !Config
       -- ^ Top-level Stack configuration.
-    , lcLoadBuildConfig :: !(Maybe AbstractResolver -> NoBuildConfigStrategy -> m BuildConfig)
+    , lcLoadBuildConfig :: !(Maybe AbstractResolver -> m BuildConfig)
         -- ^ Action to load the remaining 'BuildConfig'.
     , lcProjectRoot     :: !(Maybe (Path Abs Dir))
         -- ^ The project root directory, if in a project.
     }
-
-data NoBuildConfigStrategy
-    = ThrowException
-    | ExecStrategy
-    deriving (Show, Eq, Ord)
 
 data PackageEntry = PackageEntry
     { peExtraDepMaybe :: !(Maybe Bool)