+++ src/test/Stack/ConfigSpec.hs
@@ -60,8 +60,8 @@ spec = beforeAll setup $ afterAll teardown $ do
 
 
   describe "loadConfig" $ do
-    let loadConfig' m = runStackLoggingT m logLevel False (loadConfig mempty Nothing)
-    let loadBuildConfigRest m = runStackLoggingT m logLevel False
+    let loadConfig' m = runStackLoggingT m logLevel False False (loadConfig mempty Nothing)
+    let loadBuildConfigRest m = runStackLoggingT m logLevel False False
     -- TODO(danburton): make sure parent dirs also don't have config file
     it "works even if no config file exists" $ \T{..} -> example $ do
       _config <- loadConfig' manager
@@ -80,7 +80,7 @@ spec = beforeAll setup $ afterAll teardown $ do
       setCurrentDirectory childDir
       LoadConfig{..} <- loadConfig' manager
       BuildConfig{..} <- loadBuildConfigRest manager
-                            (lcLoadBuildConfig Nothing ThrowException)
+                            (lcLoadBuildConfig Nothing)
       bcRoot `shouldBe` parentDir
 
     it "respects the STACK_YAML env variable" $ \T{..} -> inTempDir $ do
@@ -91,7 +91,7 @@ spec = beforeAll setup $ afterAll teardown $ do
         withEnvVar "STACK_YAML" stackYamlFp $ do
           LoadConfig{..} <- loadConfig' manager
           BuildConfig{..} <- loadBuildConfigRest manager
-                                (lcLoadBuildConfig Nothing ThrowException)
+                                (lcLoadBuildConfig Nothing)
           bcRoot `shouldBe` dir
 
     it "STACK_YAML can be relative" $ \T{..} -> inTempDir $ do
@@ -104,5 +104,5 @@ spec = beforeAll setup $ afterAll teardown $ do
         withEnvVar "STACK_YAML" (toFilePath yamlRel) $ do
             LoadConfig{..} <- loadConfig' manager
             BuildConfig{..} <- loadBuildConfigRest manager
-                                (lcLoadBuildConfig Nothing ThrowException)
+                                (lcLoadBuildConfig Nothing)
             bcStackYaml `shouldBe` yamlAbs